Fix more compiler warnings

+++ examples/streaming_parser.c
@@ -30,7 +30,7 @@ void usage(void) {
 const char* key = "a secret key";
 bool key_found = false;
 
-void find_string(void* UNUSED(_ctx), cbor_data buffer, size_t len) {
+void find_string(void* UNUSED(_ctx), cbor_data buffer, uint64_t len) {
   if (key_found) {
     printf("Found the value: %.*s\n", (int)len, buffer);
     key_found = false;
